Sweet Alerts Solved
I have the pleasure to say the either the SweetAlerts from user tables and delete tasks, are totally functional.

I have implemented AJAX to the signup.js, and have added a button on task table's rows to delete each specific row, for each specific task.

maybe we could do the same for users.